# openapi.yaml
openapi: 3.1.0
info:
  title: document-pipeline
  version: 1.0.0
  description: |
    Ingests documents through a configurable OCR → clarify → classify → embed pipeline,
    and stores the results in a vector database for RAG queries.

    Documents move through pipeline stages asynchronously via jobs. Each job represents
    execution of a single stage. The frontend polls for state changes and streams live
    LLM output via Server-Sent Events.
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - authentik: []
...
paths:
...
  /api/v1/chats:
    get:
      tags: [Chat]
      summary: List chats
      description: Returns a paginated list of chat summaries ordered by most recently updated.
      operationId: listChats
      parameters:
        - $ref: '#/components/parameters/page_size'
        - name: before_id
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of chats.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedChats'
    post:
      tags: [Chat]
      summary: Create a new chat
      description: Creates a new chat with optional system prompt and RAG retrieval settings.
      operationId: createChat
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatBody'
      responses:
        '200':
          description: Newly created chat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSummary'
  /api/v1/chats/{chat_id}:
    parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Chat]
      summary: Get chat with messages
      operationId: getChat
      responses:
        '200':
          description: Chat detail including full message history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Chat]
      summary: Update chat
      description: Updates mutable chat fields. Only present fields are applied.
      operationId: patchChat
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchChatBody'
      responses:
        '200':
          description: Updated chat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSummary'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Chat]
      summary: Delete a chat and all its messages
      operationId: deleteChat
      responses:
        '204':
          description: Chat deleted.
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/chats/{chat_id}/messages:
    post:
      tags: [Chat]
      summary: Send a message and stream the RAG-augmented reply
      operationId: sendChatMessage
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageBody'
      responses:
        '200':
          description: SSE stream of token events followed by a done event.
          content:
            text/event-stream:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
components:
...
  parameters:
...
    page_size:
      name: page_size
      in: query
      required: false
      description: Maximum number of results to return (default 50, max 1000).
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
...
  responses:
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ── Schemas ───────────────────────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      description: Generic error body.
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    RagRetrieval:
      type: object
      description: RAG retrieval configuration for a chat.
      properties:
        enabled:
          type: boolean
          default: true
          description: Whether RAG retrieval is enabled for this chat.
        max_sources:
          type: integer
          default: 5
          minimum: 1
          maximum: 20
          description: Maximum number of source documents to retrieve per query.
        minimum_score:
          type: number
          format: float
          default: 0.0
          description: Minimum similarity score for retrieved sources.
    SourceDoc:
      type: object
      description: A document or series corpus chunk retrieved from the vector store as a RAG source.
      required: [score]
      properties:
        document_id:
          type: string
          format: uuid
          nullable: true
          description: UUID of the source document. Null for series corpus chunks.
        series_name:
          type: string
          nullable: true
          description: Series name for corpus chunks that span multiple documents.
        title:
          type: string
          description: Document title.
        summary:
          type: string
          default: ''
          description: Brief summary of the document's content.
        date_month:
          type: string
          default: ''
          description: YYYY-MM month label for the document.
        score:
          type: number
          format: float
          description: Vector similarity score.
    ChatMessage:
      type: object
      description: A single message turn in a chat.
      required: [id, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Unique message identifier.
        external_id:
          type: string
          nullable: true
          description: External identifier assigned by the LLM provider (e.g. Ollama message ID).
        role:
          type: string
          enum: [user, assistant]
          description: Message author.
        content:
          type: string
          description: Message text.
        sources:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/SourceDoc'
          description: Retrieved source documents (only set on assistant messages).
        created_at:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp.
    ChatSummary:
      type: object
      description: Lightweight chat representation used in list views.
      required: [id, rag_retrieval, created_at, updated_at]
      properties:
        id:
          type: string
          description: Chat UUID.
        title:
          type: string
          nullable: true
          description: Auto-derived from the first user message (first 60 chars). Null until the first message is sent.
        system_prompt:
          type: string
          description: Optional freeform context injected into every system prompt.
        rag_retrieval:
          $ref: '#/components/schemas/RagRetrieval'
        created_at:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp.
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 last-updated timestamp.
    ChatDetail:
      allOf:
        - $ref: '#/components/schemas/ChatSummary'
        - type: object
          description: Full chat data including message history.
          properties:
            messages:
              type: array
              description: All messages in the chat in chronological order.
              items:
                $ref: '#/components/schemas/ChatMessage'
    PaginatedChats:
      type: object
      description: Paginated list of chats ordered by most recently updated.
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ChatSummary'
        next_page_token:
          type: string
          nullable: true
          description: Opaque cursor; pass as `page_token` on next request. Null when no more pages.
    CreateChatBody:
      type: object
      description: Options for creating a new chat.
      properties:
        system_prompt:
          type: string
          default: ''
          description: Optional freeform context injected into every system prompt.
        rag_retrieval:
          $ref: '#/components/schemas/RagRetrieval'
    PatchChatBody:
      type: object
      description: Fields to update on an existing chat. Omitted fields are unchanged.
      properties:
        title:
          type: string
          nullable: true
          description: Override the auto-derived title.
        system_prompt:
          type: string
          nullable: true
          description: Updated system prompt. Null clears it.
        rag_retrieval:
          $ref: '#/components/schemas/RagRetrieval'
    SendMessageBody:
      type: object
      description: A new user message to send to the chat.
      required: [content]
      properties:
        content:
          type: string
          description: The user's message text.
  securitySchemes:
    authentik:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.jasonfagerberg.duckdns.org/application/o/authorize/
          tokenUrl: https://auth.jasonfagerberg.duckdns.org/application/o/token/
          refreshUrl: https://auth.jasonfagerberg.duckdns.org/application/o/token/
          scopes:
            openid: OpenID Connect identity token
            profile: User profile (name, username)
            email: User email address